Change PrefetchHelper to template class
Refactor PrefetchHelper, QueryMetadataJob, DownloadItemsJob
to work with different roots types.
With PrefetchHelper we can start query data handles for
different types. User should only specify such functions as
'query' - to request metadata needed to download data,
'filter' - to manage results received after query,
'download' - download data.

Relates-To: OLPEDGE-1900
